Refactor market data fetching and parsing functions
This commit refactors the code related to fetching and parsing market data. The `fetchMarketData` function now takes an additional parameter, `hashmap`, which is used to store the close prices for each stock symbol. The parsing logic has been moved to a new function called `parseMarketData`, which takes the JSON data, symbol, and `hashmap` as parameters. The `logger` function, which is currently commented out, will be used to display the last 100 days' closing prices for a given stock symbol.

The changes aim to improve code organization and separation of concerns, making it easier to maintain and extend the functionality in the future.